tugas

+++ src/pages/index.jsx
@@ -1,11 +1,12 @@
-import Container from 'components/Container';
-import Loading from 'components/Loading';
-import Card from 'components/Card';
-import axios from 'axios';
-import { ButtonSecondary } from 'components/Button';
-import { WithRouter } from 'utils/Navigation';
-import { useState, useEffect } from 'react';
-import { data } from 'autoprefixer';
+import Container from "components/Container";
+import Loading from "components/Loading";
+import Card from "components/Card";
+import axios from "axios";
+import { ButtonSecondary } from "components/Button";
+import { WithRouter } from "utils/Navigation";
+import { useState, useEffect } from "react";
+import { data } from "browserslist";
+import { useTitle } from "utils/hooks/useTitle.js";
 
 function App(props) {
   // cosntructor start
@@ -16,31 +17,19 @@ function App(props) {
   const [loading, setLoading] = useState(true);
   const [page, setPage] = useState(1);
   // cosntructor end
+  useTitle("MY's Theatre");
 
-  // state = {
-  //   title: '',
-  //   datas: [],
-  //   skeleton: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
-  //   loading: true,
-  //   page: 1,
-  // };
-  // pengganti componentDidMount
   useEffect(() => {
     fetchData();
     // fetchPopular()
   }, []);
 
-  // componentDidMount() {
-  //   // console.log(this.props);
-  //   this.fetchData();
-  // }
-
   // manggil api
   function fetchData() {
-    // this.setState({ loading: true });
+    // setLoading(true);
     axios
       .get(
-        `https://api.themoviedb.org/3/movie/now_playing?api_key=${process.env.REACT_APP_KEY}&language=en-US&page=${page}`
+        `https://api.themoviedb.org/3/movie/now_playing?api_key=${process.env.REACT_APP_KEY}&page=${page}`
       )
       .then((res) => {
         const { results } = res.data;
@@ -55,7 +44,7 @@ function App(props) {
         setPage(newPage);
       })
       .catch((err) => {
-        console.log(err);
+        alert(err.toString());
       })
       .finally(() => {
         setLoading(false);
@@ -71,24 +60,26 @@ function App(props) {
     // di dalam browser locar storage ada value dan key,keynya dikasih nama
 
     // favMovies
-    const getMovies = localStorage.getItem('favMovies');
+    const getMovies = localStorage.getItem("favMovies");
 
-    // if (movie.id === data.id) {
     if (getMovies) {
       const parsedMovies = JSON.parse(getMovies);
+      const favMovie = parsedMovies.find((obj) => obj.title === movie.title);
+      if (favMovie) return alert("this movie already added!");
+
       parsedMovies.push(movie);
       const temp = JSON.stringify(parsedMovies);
-      localStorage.setItem('favMovies', temp);
+      localStorage.setItem("favMovies", temp);
     } else {
       const temp = JSON.stringify([movie]);
-      localStorage.setItem('favMovies', temp);
+      localStorage.setItem("favMovies", temp);
     }
   }
 
   return (
     <Container>
-      <div className="w-full flex flex-col text-center">
-        <div className="grid grid-cols-1 md:grid-cols-5 lg:grid:cols-5 gap-5 md:mt-14 md:mx-10">
+      <div className="w-full  text-center flex justify-center">
+        <div className="grid grid-cols-1  sm:grid-cols-3 md:grid-cols-5 lg:grid:cols-5 gap-5 md:mt-14 md:mx-10">
           {loading
             ? skeleton.map((item) => <Loading key={item} />)
             : datas.map((data) => (
@@ -101,8 +92,8 @@ function App(props) {
                 />
               ))}
         </div>
-        <ButtonSecondary label="Load More" onClick={() => fetchData()} />
       </div>
+      <ButtonSecondary label="More Movies..." onClick={() => fetchData()} />
     </Container>
   );
 }
